+++ packages/frontend/src/features/metricsCatalog/components/MetricsCatalogColumns.tsx
@@ -68,7 +68,7 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
     {
         accessorKey: SpotlightTableColumns.TABLE,
         header: 'Table',
-        enableSorting: false,
+        enableSorting: true,
         enableEditing: false,
         size: 150,
         Header: ({ column }) => (