+++ packages/frontend/src/pages/AuthPopupResult.tsx
@@ -3,6 +3,7 @@ import { IconAlertCircle } from '@tabler/icons-react';
 import { type FC } from 'react';
 import { useParams } from 'react-router';
 import { useMount } from 'react-use';
+import { DocumentTitle } from '../components/common/DocumentTitle';
 import SuboptimalState from '../components/common/SuboptimalState/SuboptimalState';
 import LightdashLogo from '../components/LightdashLogo/LightdashLogo';
 
@@ -21,7 +22,7 @@ const AuthPopupResult: FC = () => {
 
     return (
         <>
-            <title>Authentication - Lightdash</title>
+            <DocumentTitle title="Authentication" />
 
             <Stack>
                 <Box mx="auto" my="lg">
@@ -56,7 +57,7 @@ export const SuccessAuthPopupResult: FC = () => {
 
     return (
         <>
-            <title>Authentication - Lightdash</title>
+            <DocumentTitle title="Authentication" />
 
             <Stack>
                 <Box mx="auto" my="lg">